Move crd package to transformerconfig.factory.
