+++ pkg/transformerconfig/constants.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package crds
+package transformerconfig
 
 // Annotation is to mark a field as annotations.
 // "x-kubernetes-annotation": ""
